Migrate structures::uimage to zerocopy

+++ src/extractors/uimage.rs
@@ -59,7 +59,7 @@ pub fn extract_uimage(
             result.size = Some(uimage_header.header_size);
 
             // Check the data CRC
-            let data_crc_valid: bool = crc32(image_data) == (uimage_header.data_checksum as u32);
+            let data_crc_valid: bool = crc32(image_data) == uimage_header.data_checksum;
 
             // If the data CRC is valid, include the size of the data in the reported size
             if data_crc_valid {
